requirements.txt
torch
pandas
numpy
pyyaml
tqdm
scikit-learn
kagglehub
streamlit

